Create bullet pool: fix bullet lagging
This patch fixes the bullet lagging, it creates the bullet pool for
every tank->ammunition and instead of destroying the bullet entity every
time, it just relocates and reuses it.
Also bullet movement moved to tank update method instead and it
guarantees that if tank update is called, bullet will also move in
frame.